moved top nav into layout
