+++ services/frontend-v3/src/components/changeLanguage/ChangeLanguage.jsx
@@ -3,12 +3,12 @@ import {} from "antd";
 
 import ChangeLanguageView from "./ChangeLanguageView";
 
-export default class SideNav extends React.Component {
-  render() {
-    return (
-      <ChangeLanguageView
-        changeLanguage={this.props.changeLanguage}
-      ></ChangeLanguageView>
-    );
-  }
+function ChangeLanguage(props) {
+  return (
+    <ChangeLanguageView
+      changeLanguage={props.changeLanguage}
+    ></ChangeLanguageView>
+  );
 }
+
+export default ChangeLanguage;
