+++ services/frontend-v3/src/components/layouts/appLayout/AppLayoutView.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { Layout } from "antd";
-import TopNav from "../../topNav/TopNav";
+import TopNav from "./topNav/TopNav";
 import SideNav from "../../sideNav/SideNav";
 
 const { Content } = Layout;
